Add didSet observer to special requests toggle
This change was needed to eliminate a bug where the options to add
extra frosting and/or sprinkles would stay enabled even if special
requests themselves are turned off. Adding the didSet observer to
the special requests toggle will make sure these are always false when
special requests are toggled.